fix(HomeScreen): blank cause by api calls

+++ core-network/src/main/java/dev/halim/core/network/response/libraryitem/FileMetadata.kt
@@ -9,7 +9,7 @@ data class FileMetadata(
   @SerialName("ext") val ext: String = "",
   @SerialName("path") val path: String = "",
   @SerialName("relPath") val relPath: String = "",
-  @SerialName("size") val size: Int = 0,
+  @SerialName("size") val size: Long = 0,
   @SerialName("mtimeMs") val mtimeMs: Long = 0,
   @SerialName("ctimeMs") val ctimeMs: Long = 0,
   @SerialName("birthtimeMs") val birthtimeMs: Long = 0,
